requirements.txt
pydub
simpleaudio
python-dotenv
homeassistant-api
xdg-base-dirs